Moved some helper files into proper folders
