+++ ds3-sdk/src/main/java/com/spectralogic/ds3client/exceptions/UnableToConvertToContentsException.java
@@ -12,10 +12,10 @@
  *   specific language governing permissions and limitations under the License.
  * ****************************************************************************
  */
+package com.spectralogic.ds3client.exceptions;
 
-package com.spectralogic.ds3client.models;
-public class UnableToConvertToCommonPrefixesException extends RuntimeException {
-    public UnableToConvertToCommonPrefixesException(final String s) {
+public class UnableToConvertToContentsException extends RuntimeException {
+    public UnableToConvertToContentsException(final String s) {
         super(s);
     }
 }
